Move `NonSkinnableScreenPlaceholder` to outer scope
